initialize_service moved to service layer
