+++ back/database/connection.py
@@ -26,22 +26,22 @@ def get_engine() -> Any:
     """Dependency function to get the SQLAlchemy engine."""
     return engine
 
-def create_book(db: Session, book_data: BookExternalInfo) -> Books:
-    """
-    Creates a new Books record from the external BookInfo model,
-    filling in default values for local-only fields.
-    """
-    db_book = Books(
-        title=book_data.title,
-        author=book_data.author,
-        isbn=book_data.isbn,
-        cover_image_url=book_data.cover_image_url,
-        description="Book details fetched from external source.", 
-        status="store",
-    )
+# def create_book(db: Session, book_data: BookExternalInfo) -> Books:
+#     """
+#     Creates a new Books record from the external BookInfo model,
+#     filling in default values for local-only fields.
+#     """
+#     db_book = Books(
+#         title=book_data.title,
+#         author=book_data.author,
+#         isbn=book_data.isbn,
+#         cover_image_url=book_data.cover_image_url,
+#         description="Book details fetched from external source.", 
+#         status="store",
+#     )
 
-    db.add(db_book)
-    db.commit()
-    db.refresh(db_book)
-    return db_book
+#     db.add(db_book)
+#     db.commit()
+#     db.refresh(db_book)
+#     return db_book
